Added RenEx.Clone project.
RenEx.Clone will be a console application that makes a copy of directory
structure for RenEx testing.
It takes snapshot of contents of a directory and creates an identical
directory trees, where each file is an empty dummy. This way a wide
range of file renaming scenarios can be arranged and reused without the
need to actually have such files.